- Libs beside, reentrance lock is better than sync

+++ src/main/kotlin/org/harryng/kotlin/demo/entity/CounterImpl.kt
@@ -1,8 +1,10 @@
 package org.harryng.kotlin.demo.entity
 
+import java.util.concurrent.atomic.AtomicLong
+
 data class CounterImpl(
     override var id: String,
-    override var value: Long
+    override var value: AtomicLong
 ) : CounterModel(id, value) {
-    constructor() : this("", 0)
+    constructor() : this("", AtomicLong(0L))
 }
